.wav header implemented to big-endian
The ESP32 itself has a little-endian architecture, but the I2S peripheral stores that in the reversed order, i.e. with a big-endian architecture.

Therefore when introducing the header to the file it will be in little-endian format and the following data in big-endian.

A possible workaround is to swap the bytes of each individual sample of stored data, but it is infeasible due to the the size of the recording.

A better solution is to embrace the big-endian architecture and swap the bytes of each information on the wav header, therefore that is what has been implemented in this commit.